Work around sporadic failures in the subprocess buffering tests.
If the test takes too long to run, some assertions fail because they
expect some lines to be output very fast, while some a bit slower. If
the machine is a bit loaded when the text is executed, all the lines
take a long time to be displayed, which causes spurious errors.

The fix is trivial: it consists of ignoring the additional lines using
doctest's ellipsis.

Unfortunately, the ellipsis is sometimes not flexible enough (you can't
start a result block with '...' and the test fails if you use the
ellipsis to omit in a place where there's nothing to omit), so you have
to play a bit around those limitations to have a stable test.

Thanks to Jonathan Ballet for the patch!