Fix tests on Python 2.6.
When I started writing this test, I intended to use TextTestResult, printing
to (doctest-captured) stdout.  Then I changed my mind but didn't change the
code.  The code worked on Python 2.7, but not on 2.6, and I didn't notice
because I was using detox, which didn't notice because setup.py test failed to
set the exit code properly.