Fix stale bytecode cleanup logic.
The problem:

  pyfile = file[:-14] + '.py'

works nicely when file is something like 'foo.cpython-33.pyc', but fails
miserably when file is 'foo.pypy-19.pyc'.  Result: all pyc files are
cleaned, even when they're not stale.

Actually, this whole __pycache__ logic is unnecessary: PEP-3147 clearly
states that sourceless pyc files are imported only when they appear in
the regular source directory.  Experiments confirm that.  This leads us
to...

The solution:

  do not recurse into __pycache__ when cleaning stale bytecode files.