Do not be silent if a layer goes missing
When we run some test layers in a subprocess (e.g. when you use
zope-testrunner -j2), each subprocess crawls the --test-path to find the
tests again.

If the --test-path is a relative pathname, and the parent process
changes the working directory during the test run, this will cause the
child process to fail to find the tests.  That is an error, and it
shouldn't be silently ignored.

This change needs automated tests!  I only tested it manually (and my
working directory has other uncommitted changes).