Fix non-determinism in testrunner-layers-buff.txt
This part of the test needs just one layer, running in a subprocess.
The -j2 is necessary because the output differs (those [Parallel ...]
sections), but we don't actually need multiple parallel layers.  We had
two to force the 2nd layer to run in a subprocess, but now the test
runner adds such a layer for us automatically, when -j is used.

BTW it's useful to see th original commit where this test was added
(https://github.com/zopefoundation/zope.testing/commit/701f9911873f45cdbfde5a3409a04b93a35f2935)
in order to understand its original purpose.