Automatically enable the default warnings
Fixes #54

The handling of the arguments and warnings is directly lifted from
``unittest``.

I introduced a new method, run_tests_with_warnings, that calls the
pre-existing run_tests. In the (unlikely?) event of subclasses that
override run_tests, this will ensure that they get the warnings by
default. It will also allow subclasses to change warning behaviour
without changing test behaviour.

This started producing a bunch of ResourceWarnings under Python 3 for
unclosed files and un-waited-for child process, and that broke a
doctest, so I fixed that in spawn_layer_in_subprocess.