Add option to check for duplicate test IDs
Add a --require-unique flag to the testrunner.  When passed, this flag
causes the testrunner to check every test ID it loads to see if it's
seen it before.  If it has, it raises an error and exits.

This was originally by Graham Binns in
https://code.launchpad.net/~gmb/zope.testing/check-for-unique-tests/+merge/110262,
in response to https://bugs.launchpad.net/launchpad/+bug/682771, but
doesn't appear to have been sent upstream.  I ported it across to
zope.testrunner and generally brought it up to date.