Refactor subprocess result output to emit bytes
This fixes the underlying problems that led to test failures on Python 3
in #23, paving the way for the reintroduction of subunit support.  It
also generally makes more logical sense to emit bytes here, as can be
seen from the way that we no longer need to decode in the Python 3 case
only to have the stream encode again.

Doctests still need to work with text, but we can handle that by
monkey-patching a replacement for `doctest._SpoofOut` on Python 3 that
makes an underlying `io.BytesIO` available.  (This effectively takes the
run-time special cases for Python 3 and isolates them in the doctest
support code, which seems like an improvement.)

`_get_output_buffer` is a simplified version of `subunit._unwrap_text`,
incorporated here to allow consistent behaviour without a hard
dependency on subunit.