Add test buffering, enabled by default for subunit
Python 2.7 added a --buffer option to the unittest runner with the
following description:

  "The standard output and standard error streams are buffered during
  the test run.  Output during a passing test is discarded.  Output is
  echoed normally on test fail or error and is added to the failure
  messages."

Add a matching option to zope.testrunner.

This is important for subunit output: tests that output extra things to
stdout may corrupt the subunit stream (especially with v1).  This option
is therefore always enabled for --subunit and --subunit-v2.

This is based loosely on work by Graham Binns, Benji York, and Brad
Crittenden in Launchpad's zope.testing fork; but I took a different
approach that's much closer to how modern unittest works, and I haven't
yet bothered trying to buffer output from layers (on the assumption that
there will be many fewer layers than tests, making it much easier to fix
them individually to be quiet).