Fix failure counting across layers
When tests in multiple layers are being run, zope.testrunner outputs a
per-layer summary after each layer, and an overall summary at the end.
The addition of support for unexpected successes in 4.0.1 caused the
per-layer summary to report incorrect failure counts: rather than
reporting the number of failures in the current layer (as for errors and
skips), it reported the cumulative number of failures so far in the
whole test run.  This was confusing, since it could result in per-layer
summaries that appeared to report failures in the layer when none
existed.

Restore the pre-4.0.1 behaviour of having the per-layer summary report
only the number of failures in that layer.