Improve standard stream handling when buffering
One of the last remaining issues in porting Launchpad to Python 3 is
that our CI infrastructure runs tests with subunit output (thereby
enabling test output buffering), and a few tests assume that on Python 3
they can use `sys.stdout.buffer` or `sys.stderr.buffer` to get an object
suitable for writing binary data to the standard streams.  These tests
currently fail because when zope.testrunner is doing test output
buffering it replaces the standard streams with instances of
`io.StringIO`.

Now, it's true that https://docs.python.org/3/library/sys.html notes
that "the standard streams may be replaced with file-like objects like
`io.StringIO` which do not support the `buffer` attribute"; however,
there are a couple of cases where it's difficult to work around this,
for example because the assumption is baked into existing libraries like
subunit itself.  Besides, I don't think that this note in the standard
library documentation *necessarily* means that zope.testrunner has to
do something inconvenient.

An alternative approach is to use a variant of `io.TextIOWrapper`, which
provides the `buffer` attribute for tests that need it, and can easily
be made to continue to provide the `getvalue` method for internal use.
This seems like a good compromise.