Clean up lots of lint and enable flake8
I added a few things to `per-file-ignores` where it seemed that making
`flake8` happy with them directly would involve making the code less
readable or idiomatic, but most of this is just an enormous pile of
whitespace changes and similar.

In practice there should be no functional change as a result of this
commit.  The only things I can think of seem extremely unlikely, and can
only affect Python 2:

 * On Python 2, changing bare `except:` to `except BaseException:` could
   conceivably break somebody relying on being able to raise an instance
   of an old-style class as an exception (but this will break on Python
   3 anyway);

 * Changing `if t is types.InstanceType:` to `if issubclass(t,
   types.InstanceType):` could perhaps change behaviour for some very
   strange code (but `TrackRefs.update` can't possibly work on Python 3
   at the moment anyway, since `types.InstanceType` doesn't exist
   there).