Exclude tests from wheel
As the wheel's primary purpose is to be installed in a virtualenv or on a
system, the tests for testrunner itself are unnecessary to include. Including
them also makes distro-tooling such as Arch Linux's namcap produce noisy
warnings regarding missing dependencies.

The tests are still included in the sdist as per the MANIFEST.in file.